5° Semestre/ES - Engenharia de Software/Backend/SaudeConnect - Backend/pyproject.toml
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "django-template"
version = "0.1.0"
[tool.poetry.dependencies]
Django = "^5.0"
python = "^3.10"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.4.0"
requests = "^2.32.3"
django-filter = "^24.3"
[tool.poetry.dev-dependencies]

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"